Stage 1 refacting modelpartition
